Day7 uc3 crud operations

+++ src/main/java/com/application/AddressBookAPP/service/AddressBookInterface.java
@@ -11,6 +11,6 @@ public interface AddressBookInterface {
     List<ContactData> getContactData();
     ContactData getContactDataById(Long contactID);
     ContactData createContactData(ContactDTO contactDTO);
-    ContactData updateContactData(int contactID, ContactDTO contactDTO);
-    void deleteContactData(int contactID);
+    ContactData updateContactData(Long contactID, ContactDTO contactDTO);
+    void deleteContactData(Long contactID);
 }
